Move market related non-store code to a new module
