+++ orko-ui/src/containers/LimitOrderContainer.js
@@ -25,7 +25,7 @@ import * as focusActions from "../store/focus/actions"
 import * as exchangesActions from "../store/exchanges/actions"
 import { isValidNumber } from "@orko-ui-common/util/numberUtils"
 import { getSelectedCoin } from "../selectors/coins"
-import exchangeService from "../services/exchanges"
+import exchangeService from "@orko-ui-market/exchanges"
 import * as errorActions from "../store/error/actions"
 import { withAuth } from "@orko-ui-auth/index"
