+++ orko-ui/tsconfig-paths.json
@@ -6,7 +6,8 @@
       "@orko-ui-common/*": ["./modules/common/*"],
       "@orko-ui-auth/*": ["./modules/auth/*"],
       "@orko-ui-socket/*": ["./modules/socket/*"],
-      "@orko-ui-log/*": ["./modules/log/*"]
+      "@orko-ui-log/*": ["./modules/log/*"],
+      "@orko-ui-market/*": ["./modules/market/*"]
     }
   }
 }
